Fix diffing of packages in layers
This is work towards #946

When ingesting a previous run's SBOM, we would like to store only
the list of packages associated with the current layer. So we will
need to weed out duplicate packages. This commit replaces the
commonly used update_master_list with a function
resolve_context_packages. This function is different from the
update_master_list function because it is provided the whole list
of layers and packages from previous runs, rather than the
layer-by-layer diffing of packages where the master list has to
be maintained throughout the execution.

This commit completes the JSON implementation of generating SBOMs
at container build time

Signed-off-by: Nisha K <nishak@vmware.com>