Fix prereqs.fs_shell variable naming
A Prereqs object has an attribute named 'fs_shell' that represents the
container filesystem shell. This variable was incorrectly referred to as
just 'shell' in a handful of files which meant the shell value was
lost and in some cases causing errors. This commit corrects the fs_shell
variable references.

Resolves #975

Signed-off-by: Rose Judge <rjudge@vmware.com>