Show package type in default report
This commit adds package type/format information to the default report.
This is a useful datapoint as it corresponds to the package manager in
base.yml that was used to inventory the layer. Sometimes there may be
two collection  methods used per layer and it will be helpful to
differentiate between the two by noting the package format.

Resolves #984

Signed-off-by: Rose Judge <rjudge@vmware.com>