Fix Scancode collection of package licenses
There was a traceback error running Tern with Scancode for multiple
report formats due to Scancode reporting licenses as a list containing a
single string or a list containing a license dictionary when Tern was
expecting just a string. This commit resolves the issue by parsing the
license list to get the license text before assigning it to a Package
object's 'pkg_license' value.

Resolves #985 #844

Signed-off-by: Rose Judge <rjudge@vmware.com>