Commit message part explained better
The part explaining how we are supposed to add
the commit message by git can be a bit confusing
for some new contributors as most people tend to
do it with github and not git.
Added few lines in the CONTRIBUTING.md file,
which explains the procedure a little better
to avoid any confusion.Also explains procedure
if the contributor needs to make changes in his
commit message.

Resolves issue #1019

Signed-off-by: Sayantani Saha <ii.sayantani.ii@gmail.com>