Fix: Ignore newlines in os_release file
Previously, we strip any newlines and split the string on "=".
However, if the string is empty then only one value is returned.
Hence we first strip any whitespace or newline characters in each
line, and then we check if the line is a non-empty string before
splitting it into key and value based on the "=" separator.

Signed-off-by: Nisha K <nishak@vmware.com>