Remove `/` from image SPDX Identifier Reference
Currently, Tern uses image_name-tag or image_name-checksum to render a
human readable ID for an image's SPDXRef in SPDX documents. This is
problematic when there is a `/` character in the image name, in which
case, the SPDX document fails to verify because the SPDXRef does not
meet the SPDX Identifier Reference requirements [1]. This commit
replaces the `/` character in an image name with a `-`.

[1] https://spdx.github.io/spdx-spec/8-annotations/#84-spdx-identifier-reference

Signed-off-by: Rose Judge <rjudge@vmware.com>