Fix: duplicate scancode files being reported
When running scancode with an empty cache, Tern was reporting the same
file twice due to a logic bug in the add_file_data() function. This fix
adds a flag to ensure that a file is only added to the layer object
file list if it does not already exist there.

Resolves #1000

Signed-off-by: Rose Judge <rjudge@vmware.com>