Add source as package and file property
The source of a file or package is the package manager or metadata
collection utility. Add this property to File and Package object.

This information can be helpful in determining the trustworthiness of
the collection method. This is also helpful when it comes to generating
OpossumUI output files as each package and file needs to be associated
with a source.

Signed-off-by: Rose Judge <rjudge@vmware.com>