Add OpossumUI output report format
This commit adds a new output format for OpossumUI[1]. The OpossumUI
format can be generated by invoking `tern report -i <image> -f
opossumui` on the command line. The format follows the OpossumUI JSON
schema and, once generated, can be used to visually represent Tern's
findings in a UI. Once a file is open in the UI, SBOM data can be
verifed or edited as necessary.

[1] https://github.com/opossum-tool/OpossumUI

Signed-off-by: Rose Judge <rjudge@vmware.com>