Fix for SPDX JSON format to match SPDX JSON schema
The SPDX JSON format was not validating against the SPDX JSON
schema[1]. In particular, there were four main validation issues
fixed with this commit:

1) creationInfo.creations was being reported as a string and the schema
expects an array.
2) packages.filesAnalyzed was being reported as the string `false`
instead of a boolean `False` value.
3) Packages were referred to using the key `fileName` instead
of `packageFileName`.
4) Packages were being reported without a `name` value (i.e. some
`name`s were reported as `null`)

This commit also changes the spdx[*] format to report  file names as the
full path of the file instead of just the file name. This is because the
SPDX spec describes the fileName value as: "The name of the file
relative to the root of the package."

[1] https://github.com/spdx/spdx-spec/blob/v2.2.1/schemas/spdx-schema.json

Resolves #1064

Signed-off-by: Rose Judge <rjudge@vmware.com>