Add DockerException when daemon not running
If the docker service and docker.socket service are disabled, a
DockerException error is thrown when Tern tries to connect using the
default socket. This commit adds an exception for the DockerException
in addition to the current requests ConnectionError exception so that
Tern exits gracefully if the docker dameon is not running instead of
printing the DockerError traceback to the console.

Resolves #1073

Signed-off-by: Rose Judge <rjudge@vmware.com>