Allow whiteout files to be accessed by fs_hash.sh
This is work towards #1082

In order to detect when a file or directory got deleted during
the container build process, we need to keep any whiteout files.
These are files in the filesystem beginning with the `.wh.` prefix.
As such, this change reverts the change for excluding files
with the prefix `.wh.`.

Since these files are now included in the filesystem, the hashing
of the filesystem needs to include files of size 0. Hence this
change also reverts a previous change to exclude finding files with
size 0. This change will increase the operation time, but we will
include changes that will try to mitigate the processing of files
that are of size 0 and are not whiteout files.

Along with these changes, fs_hash is modified to suppress the true
outcome of `command -v`, and to standardize the formatting of the
file's permissions to make post processing of this information easier.

The post processing changes in image_layer removes the usage of
regex compiling and instead just does a split of the line string based
on the standardized delimiters used in fs_hash.sh.

Additionally, we move the modification of the fs_hash.sh file
permissions, needed to run in containers on Windows, to prep.py so
that it gets invoked just once and not for every layer.

Signed-off-by: Nisha K <nishak@vmware.com>