Remove unnecessary mounting and unshare commands
This is work towards #1082

Using `mount` and `unshare` within a container image requires
the container to run with sysadmin privileges. This is considered
unsafe on secure platforms. Hence we attempt to remove any unnecessary
mounting and the unshare command. At this point, we assume that
the root user can still run Tern within the container.

After some investigating, we found that rpm OSs require a
/dev/urandom inode to be present, and does not access either sys
nor proc inodes. Hence we remove the mounting of proc, sys and
dev inodes and instead create a /dev/urandom inode within the
filesystem we are analyzing.

As we are not mounting anything in this scenario, we change the
names of the function to reflect its actual operation. We also
remove the recursive unmounting of the filesystem at the single
layer analysis stage as the filesystem is not mounted at this stage.

We then replace the unshare command with just chroot. At this time
we still use the overlay driver to mount container images with
multiple layers.

Signed-off-by: Nisha K <nishak@vmware.com>