Enable multi-layer analysis with no mount
This is work towards #1082

Since overlay mount within a container requires the container
to run with privileges, we need another way of manifesting the
state of the filesystem at a given layer. This change adds a
new function apply_layers which uses bulk copy from one folder to
another after taking care of whiteout files. This will be the new
default operation of Tern's default analysis method. The other two
drivers can be accessed using the --driver option.

To account for the switch between using the drivers and applying
layers using the bulk copy method, we introduce the prep_layers
function that will make the switch. The rest of the changes involve
replacing instances of mount_overlay_fs with prep_layers wherever
appropriate.

To enable full operation, before adding the files to the image
layer during reading of the file data, we set the whiteout status
of the file.

Signed-off-by: Nisha K <nishak@vmware.com>