Fix application for multiple cached layers
This is work towards #1082

When using the default method of overlaying the container image's
layers, in the event of a cache hit the current implementation does
not copy over the filesystem into the merge directory. As a result,
the overlaying fails. To fix this issue, we check to see if there
was a cache hit and if the driver is the default one. If this is
the case, we prep the merge directory anyway.

In some situations, the operations to build container images create
whiteout files owned by the root user. In order to delete these files
and directories, tern needs to be able to run these operations as the
root user. Hence we use the root_command utility to remove whiteout
files and directories.

Signed-off-by: Nisha K <nishak@vmware.com>