formats: Modify layer title based on manifest
This is work towards #948

The html formatting of lists allows for an expansion UI under
the title of the layer object. The title is the layer tarball's
digest. In the Docker image layout, this can be directly taken
from the "tar_file" key. But in the OCI layout, this value is taken
from the "digest" key. This commit allows for that difference and
defaults to the layer's index if neither one of those keys exist.

Signed-off-by: Nisha K <nishak@vmware.com>