Fix retrieving image digest
The config object is the incorrect method of retrieving an image
digest as the config and layers itself creates a manifest which
is hashed to get the image digest. Instead, we will use skopeo
to retireve it. For this, we introduce a function in
tern/load/skopeo to do the remote inspection, and parse out
the image digest type and digest.

Fixes #1101

Signed-off-by: Nisha K <nishak@vmware.com>