Add option to use custom log file
Some situations require the log files to be written to a different
location than the current working directory. This commit introduces
the option `-l/--log-file` to provide a different file path. To
use this option, run:

tern -l path/to/log/file

In order to enable this functionality, some global logging
variables were moved to the do_main function.

Fixes #1113

Signed-off-by: Nisha K <nishak@vmware.com>