Add SPDX TV source comment and relationship type
This commit adds a source comment text to be used when the SPDX tag
value format reports on source packages. The comment indicates that the
source package is not actually installed in the container so as not to
confuse the document consumers.

This commit also adds a "GENERATED_FROM" relationship type to
`formats.py` which can be used to describe source and binary package
relationships in SPDX tag value reports.

Works towards #1094

Signed-off-by: Rose Judge <rjudge@vmware.com>