Add source package info to SPDX Tag Value report
This commit adds the source package name and version information to SPDX
Tag Value reports when the source package metadata is available
(currently for rpm and dpkg based package managers). It reports source
packages on a per-layer basis for binary package objects that contain
corresponding source package information.

In order to represent the source information, the source package is
presented as its own package element and described using a
`GENERATED_FROM` relationship with the corresponding binary package.

In order to implement these changes we make changes to the
`get_package_spdxref()` function by outputting spdxref information for
both source and binary packages if a corresponding source package
exists for a package object. If no source package information is
available, we return an empty string for the source spdxref.
Because source package often have the same name and version as
their binary package, we also add `-src` to the source package SPDX
refs. Lastly, we add a new function, `get_source_package_block()`
that can print out a tag value package block for source packages.

Works towards #1094

Signed-off-by: Rose Judge <rjudge@vmware.com>