Add source package info to SPDX JSON report
This commit adds the source package name and version information to SPDX
JSON reports when the source package metadata is available
(currently for rpm and dpkg based package managers). It reports source
packages on a per-layer basis for binary package objects that contain
corresponding source package information.

In order to represent the source information, the source package is
presented as its own package element and described using a
`GENERATED_FROM` relationship with the corresponding binary package.

In order to implement these changes we add a new function,
`get_source_package_dict`, that can print out a SPDX JSON
package dictionary for source packages.

Resolves #1094

Signed-off-by: Rose Judge <rjudge@vmware.com>