debug: Fix argument name and provide driver
The argument name for analyzing container images was changed from
`docker_image` to `image`, but the same change was not done for
the debug option. This commit fixes that oversight.

Further, the operation for recovery now takes account of the
driver that was used to analyze the container image layers. By
default, no unmount attempts will be made.

Some of the clean-up is delegated to the `teardown` operation,
common to all sub-commands.

Fixes #1123

Signed-off-by: Nisha K <nishak@vmware.com>