Allow explicitly turning off src-tls-verify
In situations where TLS is not enabled for a registry, users
may now turn off the option to check TLS certificates and use
HTTP rather than HTTPS to pull images using skopeo.

It is advised to not use this flag for untrusted registries and
only use it for registries hosted locally to testing or debugging.

This option is also added for the `debug` sub-command.

Fixes #1121 and #1087

Signed-off-by: Nisha K <nishak@vmware.com>