Collect extension and Tern metadata per layer
Currently, if no cache file exists and you run Tern with an extension
(like Scancode), the output report will only include extension info in
the output report. This is because when an extension option is selected
on the command line, Tern only runs the extension capability and doesn't
collect any package metadata using the scripts in base.yml.

This commit changes Tern's extension behavior to run in parallel with
Tern's default metadata collection capability. Now, reports will include
extension information and package metadata collected from Tern scripts.
This change in behavior has been requested from several users as it is
more intuitive and in line with what users expect.

As a result of this change we delete the `analyze()` function which
analyzed either a container image using the default method OR passed
analysis to an external tool. Instead, we directly call
`run_extension_layer` during layer analysis for a container image.

Resolves #1109

Signed-off-by: Rose Judge <rjudge@vmware.com>