Do not add selinux xattrs to file data
selinux extended attributes only manifest themselves if selinux
is enabled on the host machine. Hence it does not make sense to
use them to uniquely identify an image layer across host machines.
This commit checks to see if 'selinux' does not exist in the
extended attribute before appending it to the xattrlist list.

Fixes #108

Signed-off-by: nisha <nisha@ctlfsh.tech>