Add subdir for tar files in the working dir
The safeguard of adding a subdir based on the first part of the
file name is necessary for docker-like image tar archives,
which don't have a folder for each layer but rather another tar.gz
file or similar. This is the case for kaniko built docker image
archives.

Without it, tar archives will overwrite the untar_dir on every new
layer analyzed. This will result in the os-release file not being
found as the os analyzer expects the files of the first layer
in the dir but it's actually the files of the last dir.

This patch is enabling a way for using Tern with kaniko images.
See https://github.com/tern-tools/tern/issues/361

Signed-off-by: Roger Lehmann <ich@roger-lehmann.de>