Fix error executing in chroot in Ubuntu container
When Tern ran in a Ubuntu container for images with an `apk` package
manager, there was an issue finding the shell which led to a failed
package metadata command using chroot.

This commit adds a workaround to make sure Tern can find a shell when
running in a Ubuntu container, specifically when the shell is symlinked
to busybox for the container being analyzed. A more detailed description
of the problem can be found in the bug report[1].

It's unclear exactly why the realpath fails to resolve to the correct
utility location when Tern runs in a Ubuntu container while analyzing
apk-based container images. For now, this commit is meant as a
workaround for this exceptional case that is blocking several users.

[1]https://github.com/tern-tools/tern/issues/1161

Resolves #1161

Signed-off-by: Rose Judge <rjudge@vmware.com>