Include license info for deb pkgs in SPDX reports
Tern must scan the copyright files to gather any type of license
information for Debian packages and uses the debian-inspector[1]
library to do this. Once scanned, Debian licenses found are stored
in the `pkg_licenses` field in the package data model (vs `pkg_license`
field for packages found using the package manager). This was causing
them not to be reported in SPDX documents.

This commit enables Tern to report `pkg_licenses` found in Debian
packages as `LicenseRefs` for both tag value and json SPDX formats.

Resolves #1188

[1] https://github.com/nexB/debian-inspector

Signed-off-by: Rose Judge <rjudge@vmware.com>