Catch all invalid license key characters
A fix was previously made to catch a few specific invalid license
characters (48e22cf). Because there can be a high number of invalid
license characters, instead of trying to replace them all, this commit
adds a try/catch statement to catch the AttributeError exception that
gets thrown by license_expression when there is an invalid license
character.

Resolves #1203

Signed-off-by: Rose Judge <rjudge@vmware.com>