Add package supplier info to package objects
This commit adds package supplier information as an attribute in the
package class for package objects. For some package managers (like
PyPI), there is only one feasible supplier (PyPI) and this value is
set as a constant string (PyPI). For others, like rpm, the
string is determined using the /etc/os-release file based on the Linux
Distro providing the packages. While this is not a perfect way to
determine the distro/distributor, it is satisfactory to satisfty the
NTIA minimum elements for the upcoming EO 14028.

We decided to use the distro as the supplier based on conversations had
on the SPDX mailing list[1].

[1] https://lists.spdx.org/g/Spdx-tech/message/4942

Works towards #1205

Signed-off-by: Rose Judge <rjudge@vmware.com>