Correctly parse and report local image names
A recent commit[1] added the ability to analyze local images by adding
the `docker-daemon:` string before the image name. While this works
from an analysis point of view, it yields the incorrect string for
the container name in Tern's reporting. As a fix, this new commit
removes the `docker-daemon` string when instantiating a new OCIImage
or DockerImage object, yielding the correct image name in the reports.

[1]https://github.com/tern-tools/tern/commit/40b981ce54578a7bd6fc5e19e8ded455fa9a98fc

Resolves #1214
Resolves #1212

Signed-off-by: Rose Judge <rjudge@vmware.com>