Invalid chars included in SPDX declared licenses
Tern was including commas and other invalid characters in SPDX license
expressions which resulted in failed validation for generated SPDX docs.
This commit removes the inappropriate characters and
replaces them with valid ones (i.e. swapping `&` for `and`) before
checking if a license expression is valid. It also updates the related
function descriptions and comments for clarity.

Fixes: #1223

Signed-off-by: Rose Judge <rjudge@vmware.com>